Add more structure for new days
