+++ Advent of Code 2019/Advent of Code 2019.csproj
@@ -6,4 +6,7 @@
         <RootNamespace>Advent_of_Code_2019</RootNamespace>
     </PropertyGroup>
 
+    <ItemGroup>
+      <Folder Include="Days\" />
+    </ItemGroup>
 </Project>
