+++ Advent of Code 2019/Program.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using AdventOfCode2019.Days;
+
+namespace AdventOfCode2019
+{
+    public class Program
+    {
+        static void Main(string[] args)
+        {
+            Console.WriteLine(Day1.Run());
+            Console.ReadLine();
+        }
+    }
+}
