+++ Advent of Code 2019/bin/Debug/netcoreapp3.0/Advent of Code 2019.runtimeconfig.dev.json
@@ -1,8 +1,8 @@
 {
   "runtimeOptions": {
     "additionalProbingPaths": [
-      "/Users/buddy.reno/.dotnet/store/|arch|/|tfm|",
-      "/Users/buddy.reno/.nuget/packages",
+      "/Users/max.veldink/.dotnet/store/|arch|/|tfm|",
+      "/Users/max.veldink/.nuget/packages",
       "/usr/local/share/dotnet/sdk/NuGetFallbackFolder"
     ]
   }
